make shops gracefully decline doing business when their owner's account doesn't exist rather than crashing the server
Menche's server suffered a severe loss of playerdata that made all of the accounts on it invalidated until the respective players logged back on; since many have not yet, attempts to shop at the mall typically end in crashing the server and rolling back the map. This should fix that.